requirements.txt
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
factory-boy==2.12.0
ipython==7.12.0
Markdown==3.2.1
psycopg2-binary==2.8.4
